requirements.txt
altair==6.0.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
dlib-bin==20.0.1
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
imutils==0.5.4
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.19.0
numpy==2.2.6
opencv-python==4.13.0.92
packaging==26.0
pandas==2.3.3
pillow==12.2.0
protobuf==7.34.1
pyarrow==23.0.1
pydeck==0.9.1
pygame==2.6.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
scipy==1.15.3
simpleaudio==1.0.4
six==1.17.0
smmap==5.0.3
streamlit==1.56.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
